Support variable default/fallback values
It's now possible to provide a fallback for variables so the parser
doesn't fail if the variable isn't set. The fallback reference is added
after the variable name and seprated by whitespace:

{{$myvariable verb.fallback}}

This example will attempt to read $myvariable, and if that variable is
undefined it will fall back on 'verb.fallback'. If the fallback is
invalid an exception is thrown.