Allow arguments to be passed to grammar reference functions
Arguments are enclosed in parentheses, and the arguments will be passed
to the grammar reference *if it's a function*. Otherwise it will be
ignored completely. The arguments are converted to a JSON array, so they
must be valid JSON and not contain any reference delimiters.

Neither the function or the arguments are eval()'ed.

Example:

    {my.object("argument")} -> evaluates to my.object("argument")