Convert to recursive grammar format
The grammar needs to be tree-structured, this is very useful for more
complex grammars like the tumblr grammar. The generator has also been
converted to Vue, as it is faster and supports attribute directives
which looks a lot cleaner in Jade templates.

Adding words to word lists and previewing has not been implemented yet.