Implement grammar variables and fix and improve designer
Variables can now be added to a grammar, and they're available with
Baba.variable.set() or through command-line arguments for exported node
executables.

The designer has been renamed and some text has been rewritten, god
knows what would be a good, descriptive title/subtitle for this
project. It's not really just a grammar designer anymore as it exports a
full generator with both a grammar and transforms, but it could be
renamed later if necessary.